+++ client/routes.js
@@ -3,18 +3,16 @@ import s from 'underscore.string';
 import { Meteor } from 'meteor/meteor';
 import { Accounts } from 'meteor/accounts-base';
 import { Tracker } from 'meteor/tracker';
-import { Blaze } from 'meteor/blaze';
 import { HTML } from 'meteor/htmljs';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 import { BlazeLayout } from 'meteor/kadira:blaze-layout';
-import { ReactiveVar } from 'meteor/reactive-var';
 import { Session } from 'meteor/session';
-import { Template } from 'meteor/templating';
 
 import { KonchatNotification } from '../app/ui';
 import { ChatSubscription } from '../app/models';
 import { roomTypes } from '../app/utils';
 import { call } from '../app/ui-utils';
+import { createTemplateForComponent } from './createTemplateForComponent';
 
 const getRoomById = mem((rid) => call('getRoomById', rid));
 
@@ -33,66 +31,6 @@ FlowRouter.goToRoomById = async (rid) => {
 
 
 BlazeLayout.setRoot('body');
-
-export const createTemplateForComponent = async (
-	component,
-	props = {},
-	// eslint-disable-next-line new-cap
-	renderContainerView = () => HTML.DIV(),
-	url,
-
-) => {
-	const name = component.displayName || component.name;
-
-	if (!name) {
-		throw new Error('the component must have a name');
-	}
-
-	if (Template[name]) {
-		Template[name].props.set(props);
-		return name;
-	}
-
-	Template[name] = new Blaze.Template(name, renderContainerView);
-
-	Template[name].props = new ReactiveVar(props);
-
-	const React = await import('react');
-	const ReactDOM = await import('react-dom');
-	const { MeteorProvider } = await import('./providers/MeteorProvider');
-
-	function TemplateComponent() {
-		return React.createElement(component, Template[name].props.get());
-	}
-
-	Template[name].onRendered(() => {
-		Template.instance().autorun((computation) => {
-			if (computation.firstRun) {
-				Template.instance().container = Template.instance().firstNode;
-			}
-
-			ReactDOM.render(
-				React.createElement(MeteorProvider, {
-					children: React.createElement(TemplateComponent),
-				}), Template.instance().firstNode);
-		});
-
-		url && Template.instance().autorun(() => {
-			const routeName = FlowRouter.getRouteName();
-			if (routeName !== url) {
-				ReactDOM.unmountComponentAtNode(Template.instance().container);
-			}
-		});
-	});
-
-	Template[name].onDestroyed(() => {
-		if (Template.instance().container) {
-			ReactDOM.unmountComponentAtNode(Template.instance().container);
-		}
-	});
-
-	return name;
-};
 
 FlowRouter.route('/', {
 	name: 'index',
@@ -246,6 +184,7 @@ FlowRouter.route('/setup-wizard/:step?', {
 });
 
 const style = 'overflow: hidden; flex: 1 1 auto; height: 1%;';
+
 FlowRouter.route('/admin/:group?', {
 	name: 'admin',
 	action: async ({ group = 'info' } = {}) => {
